moves load functions from categorySlug to utils
