+++ src/utils.js
@@ -4,11 +4,12 @@ import {
   dataByGeography,
   fetchSelectedDataForWholeBoundingBox,
   tables,
+  fetchSelectedDataForGeoType,
 } from "./model/censusdata/censusdata";
 import GeodataApiDataService from "./model/censusdata/services/geodataApiDataService";
 import config from "./config";
 import { ladLookup } from "./model/geography/geography";
-import { fetchCensusDataBreaks } from "./model/metadata/metadata";
+import { fetchCensusDataBreaks, reverseTotalCatCodeLookup } from "./model/metadata/metadata";
 import MetadataApiDataService from "./model/metadata/services/metadataApiDataService";
 import { mapZoomBBox } from "./model/geography/stores";
 
@@ -117,3 +118,23 @@ export const updateMap = (category) => {
     fetchCensusDataBreaks(new MetadataApiDataService(), category.code, tables[category.table].total, 5, "lsoa");
   }
 };
+
+export function fetchMapDataForSelectedCat(catCode, totalCode) {
+  fetchSelectedDataForGeoType(new GeodataApiDataService(), "lad", [catCode, totalCode]);
+  fetchCensusDataBreaks(new MetadataApiDataService(), catCode, totalCode, 5, "lad");
+  fetchCensusDataBreaks(new MetadataApiDataService(), catCode, totalCode, 5, "lsoa");
+}
+
+export function lazyLoadFullTableMapData(selectedCatCode, totalCode) {
+  const catCodes = get(reverseTotalCatCodeLookup)[totalCode];
+  let catCodesToFetch = [];
+  catCodes.forEach((catCode) => {
+    if (catCode.code != selectedCatCode) {
+      catCodesToFetch.push(catCode.code);
+    }
+  });
+  fetchSelectedDataForGeoType(new GeodataApiDataService(), "lad", [...catCodesToFetch, totalCode]);
+  catCodesToFetch.forEach((catCode) => {
+    fetchCensusDataBreaks(new MetadataApiDataService(), catCode, totalCode, 5, "lad");
+  });
+}
